Add Accelerometer class with data collection and JSON output
This commit adds an `Accelerometer` class that interfaces with an MPU6050 accelerometer. The class can start and stop data collection from the sensor. When data collection is started, sensor values are continuously collected every second and output as a JSON string. This includes the accelerometer and gyroscope readings, as well as a timestamp. The JSON output makes it easy to interface with other modules or external systems.